[217] renamed .essentials package to ".basics".
