+++ java/src/moonaFramework/basics/Dynamic.java
@@ -1,4 +1,4 @@
-package moonaFramework.essentials;
+package moonaFramework.basics;
 
 public interface Dynamic {
